requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
backoff==2.2.1
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.5
distro==1.9.0
ecdsa==0.19.1
fastapi==0.135.1
fastuuid==0.14.0
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
frozenlist==1.8.0
fsspec==2026.2.0
gevent==25.9.1
geventhttpclient==2.3.9
googleapis-common-protos==1.72.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langfuse==3.14.5
litellm==1.82.0
locust==2.43.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
openai==2.24.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pgvector==0.4.2
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
python-engineio==4.13.1
python-jose==3.5.0
python-socketio==5.16.1
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.2.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
sentry-sdk==2.53.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websocket-client==1.9.0
Werkzeug==3.1.7
wrapt==1.17.3
wsproto==1.3.2
yarl==1.23.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
